Move subsystems to proper package (#63)
