+++ src/main/java/frc/robot/competition/CompetitionRobotContainer.java
@@ -22,15 +22,15 @@ import frc.robot.classes.ModuleConfig;
 import frc.robot.commands.FieldOrientedDrive;
 import frc.robot.commands.FieldOrientedWithCardinal;
 import frc.robot.commands.OrbitalTarget;
+import frc.robot.subsystems.Elevator;
 import frc.robot.subsystems.Feeder;
 import frc.robot.subsystems.Intake;
 import frc.robot.subsystems.Shooter;
+import frc.robot.subsystems.Wrist;
 import frc.robot.subsystems.chassis.Chassis;
-import frc.robot.subsystems.chassis.Elevator;
 import frc.robot.subsystems.chassis.NeoModule;
 import frc.robot.subsystems.chassis.PoseEstimator;
 import frc.robot.subsystems.chassis.SwerveModule;
-import frc.robot.subsystems.chassis.Wrist;
 import org.littletonrobotics.junction.Logger;
 import org.photonvision.PhotonCamera;
